refactoring all img folder and starting new component to show profesional history
