+++ app/components/Header.tsx
@@ -7,7 +7,7 @@ import { usePathname } from "next/navigation";
 
 import { useMobileView } from '@/app/context/MobileView';
 import { montserrat, bebas } from "@/app/fonts/fonts";
-import logoHorizontal from "@/app/img/LogoHorizontal@2x.png";
+import logoHorizontal from "@/app/img/logos/LogoHorizontal@2x.png";
 
 import { VscListFlat } from "react-icons/vsc";
 import { AiFillInstagram } from "react-icons/ai";
