+++ app/components/WhatsAppIcon.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useMobileView } from "@/app/context/MobileView";
-import whatsappLogo from "@/app/img/wspLogo.png"
+import whatsappLogo from "@/app/img/icons/wspLogo.png"
 
 import Image from "next/image";
