+++ app/img/history/equipo.jpg
